requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
bcrypt==4.3.0
black==25.1.0
catboost==1.2.8
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.2
cycler==0.12.1
dnspython==2.7.0
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.5.0
gast==0.6.0
google-pasta==0.2.0
gradio==5.32.0
gradio_client==1.10.2
graphviz==0.20.3
groovy==0.1.2
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
Jinja2==3.1.6
joblib==1.5.0
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
motor==3.7.1
mpmath==1.3.0
mypy==1.16.0
mypy_extensions==1.1.0
namex==0.0.9
narwhals==1.40.0
networkx==3.4.2
numba==0.61.2
numpy==2.1.3
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.18
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
ruff==0.11.12
safehttpx==0.1.6
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
shap==0.47.2
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tabpfn==2.0.9
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.147
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
